-  Add new StateAction handlers. -  Test StateActions that trigger other game objects, not just the attached. -  Add dynamic dispatch to allow Editor to inspect 'object' types and show the appropriate field.  (Only finished for int, float, double, Vec2,Vec3,Vec4) -  Update examples with new StateActions
It still remains to be seen if we can pass variables out of the game object without expensively boxing them.  Currently the only way to send values out is to use Reflection to determine if a field exists.  Might be possible to store a direct pointer to a field when the initial FSM is setup.